Install Fix

+++ src/database/migrations/2019_01_30_094901_user_has_acl_resource.php
@@ -15,7 +15,7 @@ class UserHasAclResource extends Migration
     {
         Schema::create('users_resources', function(Blueprint $table)
         {
-            $table->integer('user_id')->unsigned();
+            $table->bigIncrements('user_id')->unsigned();
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
             $table->integer('resource_id')->unsigned();
             $table->foreign('resource_id')->references('id')->on('resources')->onDelete('cascade');
