+++ src/database/migrations/2019_01_30_094944_user_has_acl_role.php
@@ -15,7 +15,7 @@ class UserHasAclRole extends Migration
     {
         Schema::create('users_roles', function(Blueprint $table)
         {
-            $table->integer('user_id')->unsigned();
+            $table->bigIncrements('user_id')->unsigned();
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
             $table->integer('role_id')->unsigned();
             $table->foreign('role_id')->references('id')->on('roles')->onDelete('cascade');
